add: some CLI arguments and fix bugs in bootstrapping
The bug in bootstrapping is fixed and now are correct. Use
lambda_returns instead of discounted_returns. This is because one-step
TD (bootstrapping) is the mose intuitive way to solve the problem. It
can also guarantee that the ShapedArray' shape is also the same, which
allows we use Jax.JIT.